requirements.txt
cx_Freeze==6.8.4
Flask==1.1.2
flask_expects_json==1.5.0
natsort==8.0.0
opencv_python==4.5.1.48
requests==2.26.0
Pillow==8.4.0
